% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coordinates}
\alias{coordinates}
\title{Coordinates to the 'mountains' data frame}
\format{
\subsection{\code{coordinates}}{

A data frame with 1,256 rows and 3 columns:
\describe{
\item{id}{Ascending row number}
\item{lat, long}{Latitude & longitude of the mountain}
...
}
}
}
\source{
Geocoded using the R package 'tidygeocoder' with the geocoding service ArcGIS.
}
\usage{
coordinates
}
\description{
Geocoded by name and country. There may be erros. I did not check all coordinates.
The two data frames can be merged by the 'id' column.
}
\keyword{datasets}
