% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mountains}
\alias{mountains}
\title{Data on mountains above 2,000 meters from all around the world by elevation}
\format{
\subsection{\code{mountains}}{

A data frame with 1,256 rows and 5 columns:
\describe{
\item{id}{Ascending row number}
\item{mountain}{Name of the mountain}
\item{meters, feet}{Height of the mountain in meters & feet}
\item{country}{Countries in which the mountain is located}
...
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_mountains_by_elevation}
}
\usage{
mountains
}
\description{
Scraped from Wikipedia.
}
\keyword{datasets}
